save work & develop after dao layer finishes

+++ src/main/java/com/ic/er/dao/EntityMapper.java
@@ -9,7 +9,7 @@ public interface EntityMapper {
 
     List<EntityDO> selectByEntity(EntityDO entityDO);
 
-    int insert(EntityDO entityDO);
+    Long insert(EntityDO entityDO);
 
     // rarely use, please use update to change is_delete to 1
     int deleteById(Long id);
